Remove re-definition of `SomeSet`
Commit ebee42e04b13 ("A little cleanup around set types.", 2021-10-01)
added an exported definition of `SomeSet` that included the
built-in `set` type.

However, a typical Nim user is familiar with the definition of `SomeSet`
in the standard library of

    SomeSet*[A] = HashSet[A] | OrderedSet[A]

and was thus likely to read the lines

    proc parseHook*[T](s: string, i: var int, v: var SomeSet[T]) =

and

    proc dumpHook*[T](s: var string, v: SomeSet[T]) =

and believe that jsony does not support the built-in `set` type.

This commit therefore removes the `SomeSet` re-definition, which also
means that jsony once again works with code like

    import std/sets
    import pkg/jsony

    proc foo[T](s: SomeSet[T]) =
      discard

which successfully compiled with jsony 1.0.5 and earlier, but produced
an error with jsony 1.1.0 (2021-11-04):

    /tmp/foo.nim(4, 16) Error: ambiguous identifier: 'SomeSet' -- use one of the following:
      sets.SomeSet: SomeSet
      jsony.SomeSet: SomeSet

Support for the built-in `set` type was originally added by
commit 9a4f118ac677 (2021-10-01).

Fixes: #33