Added more elaborate explanation around dumpHook
I ran face first into this issue of not understanding what `s` in the example 
dump-hook was. In my naivity I *replaced* the contents of `s` with the 
serialization-string of my custom datatype.

To avoid others from running into the same issue and foster a slightly better
understanding of what happens in dumpHooks I added these two sentences.